STMP-67 code review feedback

+++ grails-app/domain/uk/co/wonderlane/wlpos/reporting/Sale.groovy
@@ -10,7 +10,7 @@ class Sale {
     int retailerId
     int storeId
     int tillId
-    Integer quantity
+    BigDecimal quantity
     BigDecimal costPrice
     BigDecimal retailPrice
     BigDecimal vatAmount
